% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{xt}
\alias{xt}
\title{Construct a two-way frequency table (matrix) from a 2 or 3
variable array of values}
\usage{
xt(trn, fr.col, to.col, cnt.col = NULL, brk = NULL)
}
\arguments{
\item{trn}{An array or data frame with two or three columns
(either from, to or from, to, count). Each row should at
least designate the from-group & to-group (count is optional).
\if{html}{
  \tabular{ccc}{
  From   \tab To   \tab Count \cr
  x1  \tab x1  \tab 40  \cr
  x1  \tab x2  \tab 89  \cr
  ... \tab ... \tab ...
  }
}
\if{latex}{
\deqn{
  \left(
    \begin{array}{cccc}
x1 & x1 & 40 & \dots \\
x1 & x2 & 89 & \dots \\
... & ... & ... & \dots \\
\vdots & \vdots & \vdots & \ddots
      \end{array}\right)
  }
}}

\item{fr.col}{The column number of \code{t} for the `from' variable, which
will be reformed to vary across columns in the output matrix.}

\item{to.col}{The column number of \code{t} for the `to' variable, which
will be reformed to vary across rows in the output
matrix.}

\item{cnt.col}{(optional) The column number of \code{t} for the `count'
variable. If unspecified then each row will be counted as one
observation.}

\item{brk}{(optional) Either a value indicating the number of bins to
aggregate \code{t} into, or a vector specifying the breakpoints to use
if probabilistically equivalent bin ranges are desired. Equivalent
ranges can be calculated using \code{\link{brkpts}}.}
}
\value{
Returns a two-way table of frequencies in the form:
\if{html}{
  \tabular{ccccc}{
       \tab     \tab Fr  \tab    \tab      \cr
       \tab 40  \tab 38  \tab 37  \tab \dots\cr
  To   \tab 89  \tab 89  \tab 87  \tab \dots\cr
       \tab 75  \tab 74  \tab 89  \tab \dots\cr
       \tab \dots \tab \dots \tab \dots \tab \dots
  }
}
\if{latex}{
\deqn{
  \left(
    \begin{array}{cccc}
40 & 38 & 37 & \dots \\
89 & 89 & 87 & \dots \\
75 & 74 & 89 & \dots \\
\vdots & \vdots & \vdots & \ddots
      \end{array}\right)
  }
}
}
\description{
Construct a two-way frequency table (matrix) by reforming and
aggregating counts into bins.
}
\details{
xt requires an array or data frame of numeric values and returns
a two-way table of frequencies that can be optionally grouped into bins
of a specified range.
}
\examples{
data(transitions)                # Load example data
b <- brkpts(transitions$phenofr, # 10 probabilistically
            10)                  #  equivalent breakpoints
m <- xt(transitions, fr.col=2,   # Construct a two-way table
        to.col=3, cnt.col=4,
        brk=b)
}
\references{
PAPER TITLE.
}
\author{
Bjorn J. Brooks, Lars Y. Pomara, Danny C. Lee
}
